% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snb_dist.r
\name{cdsnb}
\alias{cdsnb}
\title{The Conditional Stopped Negative Binomial Density}
\usage{
cdsnb(d, s, t, prior = c(0.5, 0.5))
}
\arguments{
\item{d}{a sequence of 1's and 0's corresponding to the binomial process.}

\item{s}{the top barrier for the snb process.}

\item{t}{the right barrier for the snb process.}

\item{prior}{the shape parameters of the prior on the success probability.}
}
\description{
The conditional stacked snb density function. This function gets
the distribution of the stopping time when the binomial process has not 
reached one of its endpoints. The success probability is fitted using 
fit_flips with specified prior.
}

