% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snb_dist.r
\name{cdsnb_stack_plot}
\alias{cdsnb_stack_plot}
\title{The Conditional Stopped Negative Binomial Density Plot}
\usage{
cdsnb_stack_plot(d, shape, s, t)
}
\arguments{
\item{d}{a sequence of 1's and 0's corresponding to the binomial process.}

\item{s}{the top barrier for the snb process.}

\item{t}{the right barrier for the snb process.}

\item{prior}{the shape parameters of the prior on the success probability.}
}
\description{
A plot of the stacked snb density function. The plot shows the distribution
of the stopping time when the binomial process has not reached one of 
its endpoints. The success probabilty is fitted using the fit_flips function
with specefied prior.
}

