% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snb_dist.r
\name{dsnb_stack_plot}
\alias{dsnb_stack_plot}
\title{The Stopped Negative Binomial p.m.f. Stack-Plot}
\usage{
dsnb_stack_plot(p, s, t, x, offset)
}
\arguments{
\item{p}{the probability of a success on each coin flip.}

\item{s}{the top barrier for the snb process.}

\item{t}{the right barrier for the snb process.}

\item{x}{the range of the distribution (defaults to min(s,t):(t+s-1)).}

\item{offset}{an offset on the domain of the distribution. This is 
used when getting the conditional distribution where the domain does 
not start at 1.}
}
\value{
a plot of the probability mass function.
}
\description{
The stacked plot of the probability mass function for the snb showing
the contributions from N (the top barrier) and R (the right barrier).
}

