% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snb_dist.r
\name{dsnb_stacked}
\alias{dsnb_stacked}
\title{Stack the distribution by responders and non-responders.}
\usage{
dsnb_stacked(x, p, s, t)
}
\arguments{
\item{x}{quantile}

\item{p}{success probability}

\item{s}{number of successes}

\item{t}{number of failures}
}
\description{
Stacked distribution function for the stopped negative binomial distribution.
}

