% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snb_dist.r
\name{dsnbc_stack}
\alias{dsnbc_stack}
\title{The "Stacked" Compound Negative Binomial Density Function}
\usage{
dsnbc_stack(d, s, t, shape1 = 0.5, shape2 = 0.5, x = min(s, t):(t + s -
  1))
}
\arguments{
\item{d}{the data, a vector of 0 and 1 values.}

\item{s}{the top barrier for the snb process.}

\item{t}{the right barrier for the snb process.}

\item{shape1}{the value of the first shape parameter on the prior}

\item{shape2}{the value of the second shape parameter on the prior}

\item{x}{the range of the distribution (defaults to min(s,t):(t+s-1)).}
}
\value{
the "stacked" density.
}
\description{
This function returns the "stacked" density function showing the
contribution from each of the end points to the total mass.
}

