% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snb_dist.r
\name{dsnbc_stack_plot}
\alias{dsnbc_stack_plot}
\title{Stacked Plot of the Compound Stopped Negative Binomial Density}
\usage{
dsnbc_stack_plot(d, s, t, shape1 = 0.5, shape2 = 0.5, x = min(s, t):(t + s
  - 1))
}
\arguments{
\item{d}{the data, a vector of 0 and 1 values.}

\item{s}{the top barrier for the snb process.}

\item{t}{the right barrier for the snb process.}

\item{shape1}{the value of the first shape parameter on the prior}

\item{shape2}{the value of the second shape parameter on the prior}

\item{x}{the range of the distribution (defaults to min(s,t):(t+s-1)).}
}
\value{
a plot of the probability mass function.
}
\description{
The stacked plot of the probability mass function for the snb showing
the contributions from N (the top barrier) and R (the right barrier).
}

