% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snb_dist.r
\name{esnb}
\alias{esnb}
\title{Expected value of an SNB}
\usage{
esnb(p, s, t)
}
\arguments{
\item{p}{success probability}

\item{s}{number of successes}

\item{t}{number of failures}
}
\description{
Find the expected size of an SNB distribution with specified parameters.
}

