% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snb_stat.r
\name{fit_flips}
\alias{fit_flips}
\title{Fit the Stopped Negative Binomial Distribution}
\usage{
fit_flips(x, s, t, prior = c(0.5, 0.5))
}
\arguments{
\item{x}{The vector of 1's and 0's to fit.}

\item{s}{The ceiling for the snb process.}

\item{t}{The right barrier for the snb process.}

\item{prior}{The two element vector giving the parameters for the beta prior.}
}
\description{
Fit a vector of zero's and ones according to the stopped negative
binomial distribution with specified 's' and 't' parameters and
a beta prior.
}

