% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snb_dist.r
\name{kplot}
\alias{kplot}
\title{The K-Plot for the Binomial Process}
\usage{
kplot(flips, s, t)
}
\arguments{
\item{flips}{the sequence of coin flips (1's and 0's) to visualize.}

\item{s}{the top barrier for the Bernoulli process.}

\item{t}{the right barrier for the Bernoulli process.}
}
\description{
Visualize the stopped Bernoulli process with a horizontal step axis and a 
vertical axis counting the number of successes.
}
\examples{
flips = c(0, 0, 1)
kplot(flips, 2, 3)
}

