% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snb_dist.r
\name{snb_flips}
\alias{snb_flips}
\title{Simulate the binomial process}
\usage{
snb_flips(n, prob, s, t, drop = TRUE)
}
\arguments{
\item{n}{the number of trajectories to simulate.}

\item{prob}{the probability of a head.}

\item{s}{the number of heads to stop at.}

\item{t}{the number of tails to stop at.}

\item{drop}{if TRUE then return the results as a matrix. Otherwise return as a list.}
}
\description{
Generate coin flip trajectories that stop after either s heads or t tails.
}

