% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_app.r
\name{snb_sim_app}
\alias{snb_sim_app}
\title{Run the snb Simulation Web Application}
\usage{
snb_sim_app(port = 8100L, launch.browser = getOption("shiny.launch.browser",
  interactive()), workerId = "")
}
\arguments{
\item{port}{The TCP port that the application should listen on. Defaults to port 8100.}

\item{launch.browser}{If true, the system's default web browser will be launched automatically after the app is started. Defaults to true in interactive sessions only.}

\item{workerId}{Can generally be ignored. Exists to help some editions of Shiny Server.}
}
\description{
This function launches a shiny web page for visualizing the stopped
negative binomial simulation in the browser.
}

