% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snb_dist.r
\name{vsnb}
\alias{vsnb}
\title{Variance of the SNB Distribution}
\usage{
vsnb(p, s, t)
}
\arguments{
\item{p}{success probability}

\item{s}{number of successes}

\item{t}{number of failures}
}
\description{
Find the variance of the SNB distribution with specified parameters.
}

