% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snb_dist.r
\name{zplot}
\alias{zplot}
\title{The Z-Plot for the Binomial Process}
\usage{
zplot(flips, s, t, show_arrows = TRUE, unif_jitter = 0.2, xlab = NULL,
  ylab = NULL)
}
\arguments{
\item{flips}{the sequence of coing flips (1's and 0's) to visualize.
Note that this can be a list in which case multiple processes will be 
shown.}

\item{s}{the top barrier for the Bernoulli process.}

\item{t}{the right barrier for the Bernoulli process.}

\item{show_arrows}{should arrows be shown in the Bernoullis process path?}

\item{unif_jitter}{for multiple flip paths, how much jitter to add 
(default is 0.2).}

\item{xlab}{the name of the x axis.}

\item{ylab}{the name of the y axis.}
}
\description{
Visualize the stopped Bernoulli process with horizontal axis counting 
successes and vertical axis counting failure.
}
\examples{
flips = c(0, 0, 1)
zplot(flips, 2, 3)
}

