% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snb_dist.r
\name{cdsnb_stacked}
\alias{cdsnb_stacked}
\title{The Conditional Stopped Negative Binomial Density}
\usage{
cdsnb_stacked(x, shape, s, t)
}
\arguments{
\item{x}{quantile}

\item{shape}{the shape parameters of the beta prior.}

\item{s}{the top barrier for the snb process.}

\item{t}{the right barrier for the snb process.}
}
\description{
The conditional stacked snb density function. This function gets
the distribution of the stopping time when the binomial process has not 
reached one of its endpoints. The success probability is fitted using 
fit_flips with specified prior.
}

